Upgrade of Add Anime

+++ func/add_anime.py
@@ -23,18 +23,30 @@ def add_anime(message):
             if message.reply_to_message.forum_topic_created is None:
                 chat_username = message.chat.username
                 if chat_username is not None:
-                    if chat_username == "OtakuSenpai2020":
-                        message_id = message.reply_to_message.id
-                        topic_id = message.message_thread_id if message.is_topic_message else None
-                        if topic_id and topic_id == 251973:
-                            text = f"Título: {args[1]} \nEnlace: https://t.me/{chat_username}/{topic_id}/{message_id}"
-                            msg = bot.reply_to(message, text)
-                            time.sleep(5)
-                            bot.delete_message(msg.chat.id, msg.message_id)
+                    chat_id = message.chat.id
+                    user_id = message.from_user.id
+                    chat_member = bot.get_chat_member(chat_id, user_id)
+                    if chat_member.status in ['administrator', 'creator']:
+                        if chat_username == "OtakuSenpai2020":
+                            message_id = message.reply_to_message.id
+                            topic_id = message.message_thread_id if message.is_topic_message else None
+                            if topic_id and topic_id == 251973:
+                                text = f"Título: {args[1]} \nEnlace: https://t.me/{chat_username}/{topic_id}/{message_id}"
+                                print(f"Título: {args[1]} \nEnlace: https://t.me/{chat_username}/{topic_id}/{message_id}")
+                                try:
+                                    animes.insert_one({"title": args[1], "link": f"https://t.me/{chat_username}/{topic_id}/{message_id}"})
+                                except Exception as e:
+                                     bot.reply_to(message, f"Error al registrar anime en la base de datos: {e}")
+
+                                msg = bot.reply_to(message, text)
+                                time.sleep(5)
+                                bot.delete_message(msg.chat.id, msg.message_id)
+                            else:
+                                bot.reply_to(message, "Este comando solo puede ser usado en el tópico de <a href='https://t.me/OtakuSenpai2020/251973'>Anime</a>", parse_mode="HTML")
                         else:
-                            bot.reply_to(message, "Este comando solo puede ser usado en el tópico de <a href='https://t.me/OtakuSenpai2020/251973'>Anime</a>", parse_mode="HTML")
+                            bot.reply_to(message, "Este comando solo puede ser usado en el grupo de OtakuSenpai.")
                     else:
-                        bot.reply_to(message, "Este comando solo puede ser usado en el grupo de OtakuSenpai.")
+                        bot.reply_to(message, "Solo los administradores pueden usar este comando.")
                 else:
                     bot.reply_to(message, "No se puede obtener el enlace del mensaje porque el chat no tiene un nombre de usuario.")
             else:
